FrontEnd Validation - U can't add a icecream in CurrentStock Table twice
We use ActionSheet to stop the user to select the icecream twice after adding one time in CurrentStock.

We can't repeat the icecream in CurrentStock
In this commit, we stop the user to select icecream which already exist in CurrentStock table yeah man from the user side at least.
Therefore, in upcoming commit we change from foreignkey of CurrentStock with Icecream to OneToOne at least it's gives Unique Constraint error.