Same Actionsheet used in CurrentStock Add and SoldStock Add
We change query and alot of thing in this commit to get this effect.

Actionsheet in CurrentStock - Shows those icecream list which is not available in Stock table.

Actionsheet in SoldStock - Shows those icecreams list which only available in Stock table.
In order to keep the code base same we just change the sql table columns name by ALIAS :)