IMP - Sale data comes with normal database's entry_date format.
Now we change the entry_date formatting while reading the whole data instead just modify it on that section where we need to show it, i.e. SaleStockItemContainer.tsx.
Right there we just convert the entry_date to format we want that's it.
This also make the entry_date comparision easier.